Move News code into Journal, simply move children
