+++ src/components/Journal/PostListItem.tsx
